# swagger-server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger teamgame - OpenAPI 3.0
  description: |-
    This is a sample team game Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the team game, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of teamgame, then click [here](https://editor.swagger.io/?url=https://teamgame.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load teamgame OAS 2.0` menu option!_

    Some useful links:
    - [The team game repository](https://github.com/swagger-api/swagger-teamgame)
    - [The source API definition for the team game](https://github.com/swagger-api/swagger-teamgame/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
- url: http://localhost/api/
...
paths:
  /upload:
    post:
      tags:
      - upload
      summary: uploads a file
      description: uploads a file. Will be expanded later.
      operationId: upload
      parameters:
      - name: additionalMetadata
        in: query
        description: Additional Metadata
        required: false
        style: form
        explode: true
        schema:
          type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/upload_body'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "400":
          description: unsuccessful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
      security:
      - teamgame_auth:
        - write:teams
        - read:teams
      x-swagger-router-controller: Upload
  /team:
    get:
      tags:
      - team
      summary: Gets all teams
      description: Gets all teams
      operationId: teamGET
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/team'
                x-content-type: application/json
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/team'
        "400":
          description: Invalid status value
      security:
      - teamgame_auth:
        - write:teams
        - read:teams
      x-swagger-router-controller: Team
    put:
      tags:
      - team
      summary: Update an existing team
      description: Update an existing team by Id
      operationId: updateteam
      requestBody:
        description: Update an existent team in the game
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/team'
          application/xml:
            schema:
              $ref: '#/components/schemas/team'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/team'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team'
            application/xml:
              schema:
                $ref: '#/components/schemas/team'
        "400":
          description: Invalid ID supplied
        "404":
          description: team not found
        "405":
          description: Validation exception
      security:
      - teamgame_auth:
        - write:teams
        - read:teams
      x-swagger-router-controller: Team
    post:
      tags:
      - team
      summary: Add a new team
      description: Add a new team
      operationId: addteam
      requestBody:
        description: Create a new team in the league
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/team'
          application/xml:
            schema:
              $ref: '#/components/schemas/team'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/team'
        required: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/team'
            application/xml:
              schema:
                $ref: '#/components/schemas/team'
        "405":
          description: Invalid input
      security:
      - teamgame_auth:
        - write:teams
        - read:teams
      x-swagger-router-controller: Team
...
  /player:
    post:
      tags:
      - player
      summary: Create player
      description: This can only be done by the logged in player.
      operationId: createplayer
      requestBody:
        description: Created player object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/player'
          application/xml:
            schema:
              $ref: '#/components/schemas/player'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/player'
      responses:
        default:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/player'
            application/xml:
              schema:
                $ref: '#/components/schemas/player'
      x-swagger-router-controller: Player
...
  /player/:
    get:
      tags:
      - player
      summary: Gets all players
      description: ""
      operationId: playerGET
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/player'
            application/xml:
              schema:
                $ref: '#/components/schemas/player'
        "404":
          description: player not found
      x-swagger-router-controller: Player
...
components:
  schemas:
...
    player:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: theplayer
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: James
        team_id:
          type: integer
          format: int32
          example: 3
        captain:
          type: integer
          example: 1
        player_number:
          type: string
          example: "12"
      example:
        firstName: John
        lastName: James
        player_number: "12"
        name: theplayer
        id: 10
        team_id: 3
        captain: 1
      xml:
        name: player
...
    team:
      required:
      - name
      - photoUrls
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Downpour
      example:
        name: Downpour
        id: 10
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      example:
        code: 0
        type: type
        message: message
      xml:
        name: '##default'
    upload_body:
      type: object
      properties:
        csvFile:
          type: string
          format: binary
...
  securitySchemes:
    teamgame_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://teamgame3.swagger.io/oauth/authorize
          scopes:
            write:teams: modify teams in your account
            read:teams: read your teams
    api_key:
      type: apiKey
      name: api_key
      in: header